Rename command to ls.
