+++ cmd/soroban-cli/src/commands/contract/alias/ls.rs
